code cadets/requirements.txt
fastapi
streamlit
requests
pandas
numpy
python-dotenv